epic games list download

+++ gameshub_backend/gameshub/src/main/java/com/wsiz/gameshub/dto/SteamGamesListResponseDto.java
@@ -26,7 +26,7 @@ public class SteamGamesListResponseDto {
         gameAsMap.forEach(game -> {
             SteamGameDto steamGameDto = SteamGameDto.builder()
                     .name((String) game.get("name"))
-                    .appId(String.valueOf((Integer) game.get("appid")))
+                    .appId(String.valueOf(game.get("appid")))
                     .build();
             gamesList.add(steamGameDto);
         });
